[2단계] 장바구니 기능 미션 제출합니다. - 체인저 (#362)
* docs: 요구 사항 작성

* feat: Product 객체 구현

* feat: DDL 작성 및 H2 Database 환경 변수 설정

* feat: 데이터베이스에 상품 저장

* refactor: 포매팅 적용

* test: SpringBootTest 실제 환경으로 변경

* feat: 데이터베이스 상품 조회 기능

* feat: 데이터베이스 상품 수정 기능

* feat: 데이터베이스 상품 삭제 기능

* feat: admin 상품 추가 및 조회

* feat: admin 상품 수정

* feat: admin 상품 삭제

* feat: home 화면 상품 목록 출력

* test: 상태 코드 반환 타입 수정

* feat: 상품 추가 또는 수정 시 유효성 검증

* test: 통합 테스트 작성

* test: 통합 테스트 수정

* refactor: 필요 없는 개행 제거

* refactor: Controller에 Transactional annotation 적용

* refactor: Controller 구조 변경

* test: 인수 테스트 추가

* refactor: Product Controller Response 규격 변경

* refactor: Request의 유효성 검사 로직 위치 수정

* test: Controller 테스트 종류 변경

* refactor: 없는 id의 Product 삭제 요청 시 Status 400 반환

* refactor: API endpoint 수정

* refactor: Delete method의 Response status를 No content로 수정

* refactor: Request(DTO) Update와 Save로 분리

* refactor: Service 계층 도입

* refactor: Repository 도입

* docs: 기능 구현 목록 디벨롭 및 클래스 다이어그램 추가

* test: 인수 테스트 비개발자가 읽을 수 있도록 수정

* feat: Product의 동등성 비교와 해시 함수 정의

* refactor: 입력 주소의 id가 null이지 못하게 Validation 추가

* feat: 예외 처리 로직 및 로깅 추가

* feat: http 파일 추가

* docs: 2단계 기능 구현 목록 추가

* feat: 사용자 이메일 검증 로직

* feat: 이메일 검증 로직 추가

* feat: 비밀번호 검증 로직

* feat: 사용자 개인정보 객체 구현 및 검증 로직

* feat: User 객체 구현

* feat: User에 id 필드 추가

* refactor: Transaction 설정 변경

* feat: 사용자 저장

* feat: 사용자 생성 요청의 비밀번호 검증 로직 추가

* Test: 사용자 생성

* docs: 데이터베이스 설명 수정

* feat: 등록된 사용자 정보 반환 및 선택

* feat: 장바구니 생성

* feat: 특정 ID의 상품을 장바구니에서 제거

* feat: 로그인 및 장바구니 화면 구현

* feat: 상품 1개 DB에서 ID로 조회

* feat: 존재하지 않는 ID로 상품을 조회하는 경우 처리

* feat: 장바구니에 상품 추가

* feat: 장바구니에서 상품 삭제

* refactor: Authorization 헤더 상수 스프링 라이브러리에서 사용

* test: 실패하는 테스트 수정

* refactor: 비밀번호 입력 검증을 DTO 내부에서 수행

* refactor: User 이름을 Member로 변경

* chore: Member 테이블 DDL 추가

* refactor: 바뀐 함수명 적용

* docs: 기능 구현 목록 수정

* chore: 더미 데이터 삽입 쿼리 추가

* refactor: Repository와 DAO 분리

* feat: 사용자 DB에 저장

* feat: 표시되는 정보에서 비밀번호 제외, 이름과 휴대전화 번호 추가

* feat: DB에서 모든 사용자 정보 조회

* feat: DB에서 이메일로 사용자 조회

* chore: 장바구니 DDL 작성

* feat: 장바구니에 상품 추가

* refactor: 장바구니 추가, 삭제 로직 수정

* feat: DB에서 장바구니 상품 삭제

* feat: DB에서 사용자의 모든 장바구니 상품 불러오기

* test: 장바구니 상품 DB 테스트 구조 변경

* feat: 장바구니 상품 삭제 API 연결

* docs: 클래스 다이어그램 수정, ERD 추가

* refactor: Service 메서드의 반환값을 Domain에서 DTO로 변경

* feat: 상품 추가/수정/삭제 시 로그인 필수

* feat: 모든 요청 응답 Log 기록

* feat: 로그인 중복 로직 제거

* feat: 메서드 분리

* refactor: 예외 메시지와 이름 구분하는 문자 상수화

* refactor: 필요 없는 어노테이션 제거

* refactor: URI 수정

* refactor: static 메서드 제거하고, 인스턴스 생성하여 호출하도록 변경

* refactor: 인터페이스 내부의 사용하지 않는 메서드 제거

+++ src/main/resources/static/js/admin.js
@@ -45,8 +45,18 @@ form.addEventListener('submit', (event) => {
 });
 
 const createProduct = (product) => {
+    const credentials = localStorage.getItem('credentials');
+    if (!credentials) {
+        alert('사용자 정보가 없습니다.');
+        window.location.href = '/users';
+        return;
+    }
+
     axios.request({
-        url: '/admin/product',
+        url: '/products',
+        headers: {
+            'Authorization': `Basic ${credentials}`
+        },
         method: 'POST',
         data: {
             "name": product["name"],
@@ -61,8 +71,18 @@ const createProduct = (product) => {
 };
 
 const updateProduct = (product) => {
+    const credentials = localStorage.getItem('credentials');
+    if (!credentials) {
+        alert('사용자 정보가 없습니다.');
+        window.location.href = '/users';
+        return;
+    }
+
     axios.request({
-        url: '/admin/product/' + product['id'],
+        url: '/products/' + product['id'],
+        headers: {
+            'Authorization': `Basic ${credentials}`
+        },
         method: 'PUT',
         data: {
             "name": product["name"],
@@ -77,9 +97,19 @@ const updateProduct = (product) => {
 };
 
 const deleteProduct = (id) => {
+    const credentials = localStorage.getItem('credentials');
+    if (!credentials) {
+        alert('사용자 정보가 없습니다.');
+        window.location.href = '/users';
+        return;
+    }
+
     axios.request({
-        url: '/admin/product/' + id,
+        url: '/products/' + id,
         method: 'DELETE',
+        headers: {
+            'Authorization': `Basic ${credentials}`
+        },
     }).then((response) => {
         window.location.reload();
     }).catch((error) => {
